WIP: lisp- and non-lisp-mode
The lisp mode does work for C-like languages, but non-braced blocks
aren't detected as deeper nesting.  The whitespace mode can do that
with a small modification, which this makes.  It also renames it to
prism-non-lisp-mode, and the main mode to prism-lisp-mode.  Then
prism-mode is added, with some options, which activates the proper
mode automatically.

Probably needs more testing before being merged to master.