make sure parent faces are always inherited
The face-attribute documentation says:

To ensure that the return value is always specified and absolute, use a
value of default for INHERIT; this will resolve any unspecified or
relative values by merging with the default face (which is always
completely specified).

Why they didn't just make that the default, I don't know for sure, but
maybe they didn't have inheritable faces before, so had to do it this
way to maintain backward compatibility.