Fix: EOB error in font-lock-fontify-keywords-region
* prism.el (prism-match):
(prism-match-whitespace):

Fixes #6.  With thanks to @vuori and @piknik.

I don't understand why this is necessary, but without it, this form in
`font-lock-fontify-keywords-region' sometimes causes an end-of-buffer
error where it calls (forward-char 1):

    (or (> (point) (match-beginning 0))
        (progn (forward-char 1) t))

It has taken many hours to trace and debug this problem, so whether
this is the best solution or not, it seems to work, so it shall be.