Fix: (prism-match) Search bound
This is technically a workaround, but it should be good enough for
now, since it's not really a problem.  (It's even possible that it's
the best solution, but stepping through the code and testing in the
font-lock-studio to find out isn't worth the time right now, since
this seems to fix it.)

See #27.