Fix: (prism-match-whitespace) Enclosed lists at top-level
e.g. in Python, in a top level function definition like:

  def foo(bar):

The symbol "bar" was incorrectly given the same color as the rest of
the characters on the line.

Looking at a large Python file, this doesn't appear to break anything
else, and the git log doesn't show the changed part as being a fix for
anything else, so it appears to be the correct fix to make.  But since
we don't have regression tests, I'll let this sit on v0.4-pre for a
while, rather than making the fix on the v0.3 branch.